Move compacted-history processing to remote compaction
